src/mcp_mock/mcp_expedientes/requirements.txt
# MCP Mock de Expedientes - Dependencias

# SDK oficial MCP
mcp>=1.0.0

# Validación JWT
pyjwt>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Para transporte HTTP/SSE
starlette>=0.27.0
uvicorn>=0.23.0

# Validación de datos
pydantic>=2.0.0

# Utilidades
python-dateutil>=2.8.2
